Convert to module structure for local execution
The repository is mostly organized into notebooks, with all the model
code as well as the simulation driver code in the same notebook.  This
facilitates interactive use, but makes it more difficult to run code
locally.  This commit takes the model and driver code from the current
F-35 squadron notebook, and organizes it into modules and some
additional driver Python files.